added redis interface; added express app
