+++ lib/db.js
@@ -2,6 +2,29 @@ const redis = require("redis")
 const fs = require("fs")
 const path = require("path")
 
+class fastLookup{
+  constructor(){
+    this.meta = redis.createClient({database: 1})
+    this.search = redis.createClient({database: 2}) 
+  }
+  async connect(){
+    await this.meta.connect()
+    await this.search.connect()
+  }
+  disconnect(){
+    this.meta.quit()
+    this.search.quit()
+  }
+  async findPattern(pattern){
+    //return await this.search.KEYS(pattern)
+    return await this.search.keys(pattern)
+  }
+  async map2Meta(id){
+    return await this.meta.HGETALL(id)
+  }
+}
+
+
 class fastEntry{
   constructor(client){
     this.client = client
@@ -64,5 +87,6 @@ class fastEntry{
 }
 
 module.exports = {
-  fastEntry
+  fastEntry,
+  fastLookup
 }
