+++ utils/index.js
@@ -1,6 +1,8 @@
 const { randomSample, randInt } = require("./random")
-
+const { successify, errorify } = require("./response")
 module.exports = {
   randomSample,
-  randInt
+  randInt,
+  successify,
+  errorify
 }
