+++ utils/response.js
@@ -1,0 +1,7 @@
+exports.successify = (payload) => {
+  return { success: true, error: false, ...payload}
+}
+
+exports.errorify = (payload) => {
+  return { success: false, ...payload}
+}
